+++ app/angular/src/server/framework-preset-angular-cli.ts
@@ -114,12 +114,11 @@ async function getBuilderOptions(
    */
   const builderOptions = {
     ...browserTargetOptions,
-    ...options.angularBuilderOptions,
+    ...(options.angularBuilderOptions as JsonObject),
     tsConfig:
       options.tsConfig ??
-      options.angularBuilderOptions?.tsConfig ??
-      browserTargetOptions.tsConfig ??
-      findUpSync('tsconfig.json', { cwd: options.configDir }),
+      findUpSync('tsconfig.json', { cwd: options.configDir }) ??
+      browserTargetOptions.tsConfig,
   };
   logger.info(`=> Using angular project with "tsConfig:${builderOptions.tsConfig}"`);